pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vantage"
version = "0.1.0"
description = "A modular and extensible Python library for building AI agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Saqlain", email = "saqlain2204@gmail.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.1",
    "rich>=13.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "types-PyYAML>=0.1.0",  # For mypy type checking support
]

[project.urls]
Homepage = "https://github.com/saqlain2204/vantage"
Issues = "https://github.com/saqlain2204/vantage/issues"

[project.scripts]
vantage = "vantage.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/vantage"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
